Update README.md
Facial expressions play a crucial role in nonverbal communication, and the impact of face masks on emotion recognition is a growing concern. This study utilizes the masked AffecNet and LFW datasets to evaluate the performance of General CNN, ResNet50V2, and InceptionV3 architectures in conveying emotions. There are eight emotions in AffecNet. Meanwhile, LFW only contains three emotions. The findings reveal that ResNet50V2 achieves the highest accuracy, with approximately 74% on the LFW dataset and 73% on the AffecNet dataset. These results highlight the effectiveness of ResNet50V2 in accurately recognizing emotions, even in the presence of face masks. The outcomes of this research contribute to the field of emotion recognition and suggest potential avenues for future work, such as enhancing models and exploring additional datasets.